Move into participants groups component
