+++ src/Server.UI/Pages/Workspaces/Participants/Components/ParticipantGroups.razor
@@ -1,0 +1,34 @@
+@using Cfo.Cats.Application.Features.Participants.DTOs
+@using Cfo.Cats.Application.Features.Participants.Specifications
+
+<MudExpansionPanels MultiExpansion="true" Class="mt-4">
+    @foreach (var group in _groups)
+    {
+        <MudExpansionPanel Expanded="false" ExpandedChanged="@(expanded => OnGroupExpanded(group, expanded))">
+            <TitleContent>
+                <MudText Typo="Typo.subtitle1">
+                    <strong>@(string.IsNullOrWhiteSpace(group.Label) ? "Unassigned" : group.Label)</strong>
+                    &nbsp;(@group.Count)
+                </MudText>
+            </TitleContent>
+            <ChildContent>
+                @if (group.Loading)
+                {
+                    <MudProgressLinear Indeterminate="true" Color="Color.Primary" Class="my-2" />
+                }
+                else if (group.Items is { Length: > 0 })
+                {
+                    <ParticipantResultsView Items="group.Items"
+                                            Tabular="false"
+                                            CanReassign="CanReassign"
+                                            RecentAction="Filter.RecentAction"
+                                            @bind-SelectedIds="SelectedIds" />
+                }
+                else
+                {
+                    <MudAlert Severity="Severity.Info" Dense="true">No participants in this group.</MudAlert>
+                }
+            </ChildContent>
+        </MudExpansionPanel>
+    }
+</MudExpansionPanels>
